flask/app/requirements.txt
aioflask==0.4.0
asgiref==3.7.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==2.1.3
Flask-Cors==4.0.0
greenlet==3.0.3
greenletio==0.11.0
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.27.1
Werkzeug==2.2.2
zipp==3.17.0
